+++ src/main/java/com/j3d/engine/react/events/payloads/ReactivePointInvalidatedPayload.java
@@ -1,0 +1,10 @@
+package com.j3d.engine.react.events.payloads;
+
+import com.j3d.engine.react.events.EventPayload;
+import com.j3d.gen.grid.ReactivePoint;
+
+public class ReactivePointInvalidatedPayload extends EventPayload<ReactivePoint> {
+    public ReactivePointInvalidatedPayload(ReactivePoint e) {
+        super(e);
+    }
+}