+++ src/main/java/com/j3d/ui/engine/floating/grid2d/Grid2DPanel.java
@@ -5,7 +5,6 @@
 package com.j3d.ui.engine.floating.grid2d;
 
 import com.j3d.StaticRefs;
-import com.j3d.engine.math.convert.Conversion;
 import com.j3d.engine.math.Dim;
 import com.j3d.engine.math.ScreenPoint;
 import com.j3d.engine.math.CartesianPoint;
@@ -16,18 +15,23 @@ import com.j3d.engine.scene.find.FindResult;
 import com.j3d.engine.scene.find.Finder;
 import com.j3d.engine.scene.nodes.Thing;
 import com.j3d.engine.scene.nodes.geometry.GObject;
-import com.j3d.gen.grid.GridManager;
-import com.j3d.gen.grid.GridObject;
-import com.j3d.gen.grid.Line;
+import com.j3d.gen.grid.*;
 import com.j3d.gen.grid.Point;
 import com.j3d.ui.engine.FloatingPanel;
 import com.j3d.ui.theme.J3DTheme;
+import com.j3d.utility.generators.JLabelRichText;
+import com.j3d.utility.generic.func.QuadFunction;
 import com.j3d.utility.generic.tuple.MutablePair;
+import com.j3d.utility.generic.tuple.Triple;
 
 import javax.swing.*;
+import javax.swing.event.ChangeListener;
 import java.awt.*;
 import java.awt.event.*;
 import java.util.*;
+import java.util.function.BiFunction;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
 
 import static com.j3d.StaticRefs.*;
 
@@ -36,8 +40,10 @@ import static com.j3d.StaticRefs.*;
  * @author yetnt
  */
 public class Grid2DPanel extends javax.swing.JPanel {
-    
-    public FloatingPanel floatingPanel = new FloatingPanel("History Panel");
+
+    // TODO: JSlider doing nothing right now, write later.
+
+    public FloatingPanel floatingPanel = new FloatingPanel("Grid2d!");
     public static CartesianPoint mousePosInPanel = new CartesianPoint(0, 0);
     public static MutablePair<Integer, Integer> offset = new MutablePair<>(0, 0);
 
@@ -69,20 +75,25 @@ public class Grid2DPanel extends javax.swing.JPanel {
 
         gm = new GridManager(this);
 
+//        discoverExisting();
+        handleSpinners();
         theme();
     }
 
     private CartesianPoint toPoint(ScreenPoint p, boolean round) {
         Dim dim = getGrid().sizeDim();
         CartesianPoint cp = p.toPoint(new ConversionWithOffset(scale, dim, GridManager.fromMut()));
-        if (round) {
-            return new CartesianPoint(
-                    Math.round(cp.x),
-                    Math.round(cp.y)
-            );
-        }
+        if (round)
+            return roundPoint(cp);
 
         return cp;
+    }
+
+    public CartesianPoint roundPoint(CartesianPoint cp) {
+        return new CartesianPoint(
+                Math.round(cp.x),
+                Math.round(cp.y)
+        );
     }
 
     UUID drag = UUID.randomUUID();
@@ -293,20 +304,82 @@ public class Grid2DPanel extends javax.swing.JPanel {
         return v;
     }
 
+    public void discoverExisting() {
+        gm.invalidateAndRemoveAllReactive();
+        gm.addExistingPoints();
+    }
+
+    private void log(String string) {
+        logLbl.setText(new JLabelRichText(string).bold().underline().wrapHTML());
+    }
+
     public void theme() {
-        J3DTheme.commitAsGenericLbl(propertiesLbl, false);
-        J3DTheme.commitAsGenericLbl(jLabel1, false);
         J3DTheme.commitAsGenericLbl(renderBtn, true);
         J3DTheme.commitAsGenericLbl(xCompBtn, true);
         J3DTheme.commitAsGenericLbl(yCompBtn, true);
         J3DTheme.commitAsGenericLbl(setOriginBtn, true);
         J3DTheme.commitAsGenericLbl(presetComboBox, true);
-        J3DTheme.commitAsGenericLbl(jLabel2, false);
+        J3DTheme.commitAsGenericLbl(normalizeV1Btn, true);
+        J3DTheme.commitAsGenericLbl(normalizeV2Btn, true);
         J3DTheme.commitAsGenericLbl(jSeparator1, false);
+        J3DTheme.commitAsGenericLbl(logLbl, false);
         J3DTheme.commitAsGenericUi(btnPanel);
         J3DTheme.commitAsGenericUi(drawPanel);
         J3DTheme.commitAsGenericUi(floatingPanel);
         J3DTheme.commitAsGenericLbl(jCheckBox1, true);
+    }
+
+    void setOrigin(Vector3 origin) {
+        this.origin = origin;
+    }
+
+    void setV1(Vector3 v1) {
+        this.v1 = v1;
+    }
+
+    void setV2(Vector3 v2) {
+        this.v2 = v2;
+    }
+
+    enum SpinnerIdentifier {
+        ORIGIN, V1, V2
+    }
+
+    private HashMap<SpinnerIdentifier, Triple<JSpinner>> spinnerMap = new HashMap<>();
+
+    private void handleSpinners() {
+        QuadFunction<JSpinner, Supplier<Vector3>, BiFunction<Vector3, Double, Vector3>, Consumer<Vector3>, ChangeListener> bi = (sp, v3, applier, setter) -> (ChangeListener) e -> {
+            double v = (double) sp.getValue();
+            setter.accept(applier.apply(v3.get(), v));
+            gm.invalidateAndRemoveAllReactive();
+        };
+
+        originXSpinner.addChangeListener(bi.apply(originXSpinner, this::getOrigin, Vector3::setXComponent, this::setOrigin));
+        originYSpinner.addChangeListener(bi.apply(originYSpinner, this::getOrigin, Vector3::setYComponent, this::setOrigin));
+        originZSpinner.addChangeListener(bi.apply(originZSpinner, this::getOrigin, Vector3::setZComponent, this::setOrigin));
+
+        v1XSpinner.addChangeListener(bi.apply(v1XSpinner, this::getV1, Vector3::setXComponent, this::setV1));
+        v1YSpinner.addChangeListener(bi.apply(v1YSpinner, this::getV1, Vector3::setYComponent, this::setV1));
+        v1ZSpinner.addChangeListener(bi.apply(v1ZSpinner, this::getV1, Vector3::setZComponent, this::setV1));
+
+        v2XSpinner.addChangeListener(bi.apply(v2XSpinner, this::getV2, Vector3::setXComponent, this::setV2));
+        v2YSpinner.addChangeListener(bi.apply(v2YSpinner, this::getV2, Vector3::setYComponent, this::setV2));
+        v2ZSpinner.addChangeListener(bi.apply(v2ZSpinner, this::getV2, Vector3::setZComponent, this::setV2));
+
+        spinnerMap.put(SpinnerIdentifier.ORIGIN, new Triple<>(originXSpinner, originYSpinner, originZSpinner));
+        spinnerMap.put(SpinnerIdentifier.V1, new Triple<>(v1XSpinner, v1YSpinner, v1ZSpinner));
+        spinnerMap.put(SpinnerIdentifier.V2, new Triple<>(v2XSpinner, v2YSpinner, v2ZSpinner));
+    }
+
+    private void setSpinnerState(SpinnerIdentifier spinnerIdentifier, Vector3 v) {
+        Triple<JSpinner> sp = spinnerMap.get(spinnerIdentifier);
+        if (sp == null) return;
+
+        sp.v1().setValue(v.getX());
+        sp.v2().setValue(v.getY());
+        sp.v3().setValue(v.getZ());
+
+        log("Set " + spinnerIdentifier.name().toLowerCase() + " vector to " + v.toCommandPaletteString() + ".");
     }
 
     /**
@@ -320,15 +393,26 @@ public class Grid2DPanel extends javax.swing.JPanel {
 
         btnPanel = new javax.swing.JPanel();
         presetComboBox = new javax.swing.JComboBox<>();
-        jLabel1 = new javax.swing.JLabel();
         renderBtn = new javax.swing.JButton();
-        propertiesLbl = new javax.swing.JLabel();
         xCompBtn = new javax.swing.JButton();
         yCompBtn = new javax.swing.JButton();
-        jLabel2 = new javax.swing.JLabel();
         jSeparator1 = new javax.swing.JSeparator();
         setOriginBtn = new javax.swing.JButton();
         jCheckBox1 = new javax.swing.JCheckBox();
+        v2ZSpinner = new javax.swing.JSpinner();
+        v2XSpinner = new javax.swing.JSpinner();
+        v2YSpinner = new javax.swing.JSpinner();
+        v1ZSpinner = new javax.swing.JSpinner();
+        v1YSpinner = new javax.swing.JSpinner();
+        v1XSpinner = new javax.swing.JSpinner();
+        originZSpinner = new javax.swing.JSpinner();
+        originYSpinner = new javax.swing.JSpinner();
+        originXSpinner = new javax.swing.JSpinner();
+        directionAxesLengthSlider = new javax.swing.JSlider();
+        queryBtn = new javax.swing.JButton();
+        logLbl = new javax.swing.JLabel();
+        normalizeV1Btn = new javax.swing.JButton();
+        normalizeV2Btn = new javax.swing.JButton();
         drawPanel = new Grid();
 
         setLayout(new java.awt.BorderLayout());
@@ -338,30 +422,26 @@ public class Grid2DPanel extends javax.swing.JPanel {
         presetComboBox.setBackground(J3DTheme.BACKGROUND.color());
         presetComboBox.setForeground(J3DTheme.TEXT_PRIMARY.color());
         presetComboBox.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "None", "XY", "XZ", "YZ" }));
+        presetComboBox.setToolTipText("Allows setting of default planes");
         presetComboBox.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 presetComboBoxActionPerformed(evt);
             }
         });
 
-        jLabel1.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        jLabel1.setText("Presets");
-
         renderBtn.setBackground(J3DTheme.BACKGROUND.color());
         renderBtn.setForeground(J3DTheme.TEXT_PRIMARY.color());
         renderBtn.setText("Render");
+        renderBtn.setToolTipText("Renders all construction into the 3D viewport.");
         renderBtn.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 renderBtnActionPerformed(evt);
             }
         });
 
-        propertiesLbl.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        propertiesLbl.setText("Grid Properties Label (like obj count)");
-
         xCompBtn.setBackground(J3DTheme.BACKGROUND.color());
         xCompBtn.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        xCompBtn.setText("Change X component");
+        xCompBtn.setText("Set v1");
         xCompBtn.setToolTipText("Change the vector of the 2d plane. Rather use presets.");
         xCompBtn.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
@@ -371,7 +451,7 @@ public class Grid2DPanel extends javax.swing.JPanel {
 
         yCompBtn.setBackground(J3DTheme.BACKGROUND.color());
         yCompBtn.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        yCompBtn.setText("Change Y component");
+        yCompBtn.setText("Set v2");
         yCompBtn.setToolTipText("Change the vector of the 2d plane. Rather use presets.");
         yCompBtn.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
@@ -379,16 +459,12 @@ public class Grid2DPanel extends javax.swing.JPanel {
             }
         });
 
-        jLabel2.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        jLabel2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel2.setText("Grid Definition");
-
         jSeparator1.setBackground(new java.awt.Color(0, 0, 0));
         jSeparator1.setOpaque(true);
 
         setOriginBtn.setBackground(J3DTheme.BACKGROUND.color());
         setOriginBtn.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        setOriginBtn.setText("Set Grid Origin");
+        setOriginBtn.setText("Set Origin");
         setOriginBtn.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 setOriginBtnActionPerformed(evt);
@@ -396,10 +472,76 @@ public class Grid2DPanel extends javax.swing.JPanel {
         });
 
         jCheckBox1.setForeground(J3DTheme.TEXT_PRIMARY.color());
-        jCheckBox1.setText("delete mode");
+        jCheckBox1.setText("Delete");
+        jCheckBox1.setToolTipText("Enables delete mode to delete construction.");
         jCheckBox1.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jCheckBox1ActionPerformed(evt);
+            }
+        });
+
+        v2ZSpinner.setModel(new javax.swing.SpinnerNumberModel(1.0d, null, null, 0.5d));
+        v2ZSpinner.setToolTipText("Change the z component of the v2 vector");
+
+        v2XSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        v2XSpinner.setToolTipText("Change the x component of the v2 vector");
+
+        v2YSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        v2YSpinner.setToolTipText("Change the y component of the v2 vector");
+
+        v1ZSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        v1ZSpinner.setToolTipText("Change the z component of the v1 vector");
+
+        v1YSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        v1YSpinner.setToolTipText("Change the y component of the v1 vector");
+
+        v1XSpinner.setModel(new javax.swing.SpinnerNumberModel(1.0d, null, null, 0.5d));
+        v1XSpinner.setToolTipText("Change the x component of the v1 vector");
+
+        originZSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        originZSpinner.setToolTipText("Change the z component of the origin of the plane.");
+
+        originYSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        originYSpinner.setToolTipText("Change the y component of the origin of the plane.");
+
+        originXSpinner.setModel(new javax.swing.SpinnerNumberModel(0.0d, null, null, 0.5d));
+        originXSpinner.setToolTipText("Change the x component of the origin of the plane.");
+
+        directionAxesLengthSlider.setMaximum(20);
+        directionAxesLengthSlider.setMinimum(1);
+        directionAxesLengthSlider.setToolTipText("Changes the length of the annotative 3D direction vector arrow heads");
+        directionAxesLengthSlider.setValue(10);
+
+        queryBtn.setBackground(J3DTheme.BACKGROUND.color());
+        queryBtn.setForeground(J3DTheme.TEXT_PRIMARY.color());
+        queryBtn.setMnemonic('Q');
+        queryBtn.setText("Query");
+        queryBtn.setToolTipText("Queries the scene to display points which lie within the working plane.");
+        queryBtn.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                queryBtnActionPerformed(evt);
+            }
+        });
+
+        logLbl.setForeground(J3DTheme.TEXT_PRIMARY.color());
+        logLbl.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        logLbl.setText("logs here");
+
+        normalizeV1Btn.setBackground(J3DTheme.BACKGROUND.color());
+        normalizeV1Btn.setForeground(J3DTheme.TEXT_PRIMARY.color());
+        normalizeV1Btn.setText("norm");
+        normalizeV1Btn.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                normalizeV1BtnActionPerformed(evt);
+            }
+        });
+
+        normalizeV2Btn.setBackground(J3DTheme.BACKGROUND.color());
+        normalizeV2Btn.setForeground(J3DTheme.TEXT_PRIMARY.color());
+        normalizeV2Btn.setText("norm");
+        normalizeV2Btn.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                normalizeV2BtnActionPerformed(evt);
             }
         });
 
@@ -409,60 +551,94 @@ public class Grid2DPanel extends javax.swing.JPanel {
             btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(btnPanelLayout.createSequentialGroup()
                 .addContainerGap()
-                .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addGroup(btnPanelLayout.createSequentialGroup()
-                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(xCompBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addComponent(yCompBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(presetComboBox, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 72, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                    .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 13, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(logLbl, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addGroup(btnPanelLayout.createSequentialGroup()
-                        .addGap(3, 3, 3)
-                        .addComponent(jCheckBox1)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                         .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(renderBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addComponent(propertiesLbl, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .addGroup(btnPanelLayout.createSequentialGroup()
+                            .addGroup(btnPanelLayout.createSequentialGroup()
+                                .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                    .addComponent(xCompBtn, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 1, Short.MAX_VALUE)
+                                    .addComponent(yCompBtn, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 68, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addComponent(normalizeV1Btn, javax.swing.GroupLayout.PREFERRED_SIZE, 1, Short.MAX_VALUE)
+                                    .addComponent(normalizeV2Btn, javax.swing.GroupLayout.PREFERRED_SIZE, 1, Short.MAX_VALUE)))
+                            .addComponent(setOriginBtn, javax.swing.GroupLayout.DEFAULT_SIZE, 132, Short.MAX_VALUE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(btnPanelLayout.createSequentialGroup()
+                                .addComponent(v2XSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(19, 19, 19)
+                                .addComponent(v2YSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(18, 18, 18)
+                                .addComponent(v2ZSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(btnPanelLayout.createSequentialGroup()
+                                .addComponent(v1XSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(19, 19, 19)
+                                .addComponent(v1YSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(18, 18, 18)
+                                .addComponent(v1ZSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(btnPanelLayout.createSequentialGroup()
+                                .addComponent(originXSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(19, 19, 19)
+                                .addComponent(originYSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(18, 18, 18)
+                                .addComponent(originZSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, 49, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 13, javax.swing.GroupLayout.PREFERRED_SIZE)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(setOriginBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addContainerGap())))
+                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(btnPanelLayout.createSequentialGroup()
+                                .addComponent(jCheckBox1)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(presetComboBox, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(queryBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                            .addComponent(directionAxesLengthSlider, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addComponent(renderBtn, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addGap(17, 17, 17)))
+                .addContainerGap())
         );
         btnPanelLayout.setVerticalGroup(
             btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, btnPanelLayout.createSequentialGroup()
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+            .addGroup(btnPanelLayout.createSequentialGroup()
+                .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                     .addGroup(btnPanelLayout.createSequentialGroup()
-                        .addComponent(setOriginBtn)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(propertiesLbl, javax.swing.GroupLayout.PREFERRED_SIZE, 34, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(renderBtn)
-                            .addComponent(jCheckBox1)))
-                    .addGroup(btnPanelLayout.createSequentialGroup()
-                        .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel1)
-                            .addComponent(xCompBtn))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                         .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                             .addComponent(presetComboBox, javax.swing.GroupLayout.PREFERRED_SIZE, 22, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(yCompBtn)))
-                    .addComponent(jSeparator1))
-                .addContainerGap())
+                            .addComponent(jCheckBox1)
+                            .addComponent(queryBtn))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(directionAxesLengthSlider, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(8, 8, 8)
+                        .addComponent(renderBtn))
+                    .addGroup(btnPanelLayout.createSequentialGroup()
+                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(setOriginBtn)
+                            .addComponent(originZSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(originYSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(originXSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(xCompBtn)
+                            .addComponent(v1ZSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(v1YSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(v1XSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(normalizeV1Btn))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(btnPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(yCompBtn)
+                            .addComponent(v2ZSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(v2YSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(v2XSpinner, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(normalizeV2Btn)))
+                    .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 92, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(logLbl)
+                .addGap(0, 6, Short.MAX_VALUE))
         );
 
-        add(btnPanel, java.awt.BorderLayout.NORTH);
+        add(btnPanel, java.awt.BorderLayout.PAGE_START);
 
         drawPanel.setBackground(J3DTheme.UI_SURFACE.color());
         drawPanel.setMinimumSize(new java.awt.Dimension(496, 397));
@@ -472,7 +648,7 @@ public class Grid2DPanel extends javax.swing.JPanel {
         drawPanel.setLayout(drawPanelLayout);
         drawPanelLayout.setHorizontalGroup(
             drawPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 553, Short.MAX_VALUE)
+            .addGap(0, 596, Short.MAX_VALUE)
         );
         drawPanelLayout.setVerticalGroup(
             drawPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -485,18 +661,21 @@ public class Grid2DPanel extends javax.swing.JPanel {
     private void setOriginBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_setOriginBtnActionPerformed
         Vector3 v = ask("Input the origin Vector3 e.g. (0, 0, 0)");
         if (v == null) return;
+        setSpinnerState(SpinnerIdentifier.ORIGIN, v);
         origin = v;
     }//GEN-LAST:event_setOriginBtnActionPerformed
 
     private void xCompBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_xCompBtnActionPerformed
         Vector3 v = ask("Input v1 plane vector Vector3 e.g. (0, 0, 0)");
         if (v == null) return;
+        setSpinnerState(SpinnerIdentifier.V1, v);
         v1 = v;
     }//GEN-LAST:event_xCompBtnActionPerformed
 
     private void yCompBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_yCompBtnActionPerformed
         Vector3 v = ask("Input v2 plane vector Vector3 e.g. (0, 0, 0)");
         if (v == null) return;
+        setSpinnerState(SpinnerIdentifier.V2, v);
         v2 = v;
     }//GEN-LAST:event_yCompBtnActionPerformed
 
@@ -522,19 +701,39 @@ public class Grid2DPanel extends javax.swing.JPanel {
                     new AxisPlane(origin, v1, v2),
                     new ArrayList<>(objects)
             );
-            if (go == null) continue;
+            if (go == null) {
+                continue;
+            }
             objects.add(go);
         }
 
-        // find all stuff named render within the entire thing
-        ArrayList<FindResult> result = getSceneManager()
-                .finder().find(Thing.class, Finder.nameQuery(), "render");
+        Thing thing = null;
 
-        // create a new thing.
-        Thing thing = new Thing(
-                StaticRefs.getSceneManager().usableLayer(),
-                "render" + result.size()
-        );
+        if (gridObjects1.stream().anyMatch(g -> g instanceof ReactivePoint)) {
+            // prefer we just take the thing one of the points already is in.
+
+            thing = StaticRefs.getSceneManager().findObjectParent(
+                    gridObjects1.stream()
+                            .filter(g -> g instanceof ReactivePoint)
+                            .map(ReactivePoint.class::cast)
+                            .map(ReactivePoint::getGPoint)
+                            .findFirst()
+                            .get()
+            );
+        }
+
+        if (thing == null) {
+
+            // find all stuff named render within the entire thing
+            ArrayList<FindResult> result = getSceneManager()
+                    .finder().find(Thing.class, Finder.nameQuery(), "render");
+
+            // create a new thing.
+            thing = new Thing(
+                    StaticRefs.getSceneManager().usableLayer(),
+                    "render" + result.size()
+            );
+        }
 
         thing.addObjs(
                 objects.toArray(GObject[]::new)
@@ -550,15 +749,23 @@ public class Grid2DPanel extends javax.swing.JPanel {
         // XY XZ YZ
         switch (item) {
             case "XY" -> {
+                setSpinnerState(SpinnerIdentifier.V1, Vector3.X);
+                setSpinnerState(SpinnerIdentifier.V2, Vector3.Y);
                 v1 = Vector3.X; v2 = Vector3.Y; // same as default.
             }
             case "XZ" -> {
+                setSpinnerState(SpinnerIdentifier.V1, Vector3.X);
+                setSpinnerState(SpinnerIdentifier.V2, Vector3.Z);
                 v1 = Vector3.X; v2 = Vector3.Z;
             }
             case "YZ" -> {
+                setSpinnerState(SpinnerIdentifier.V1, Vector3.Y);
+                setSpinnerState(SpinnerIdentifier.V2, Vector3.Z);
                 v1 = Vector3.Y; v2 = Vector3.Z;
             }
             default ->  {
+                setSpinnerState(SpinnerIdentifier.V1, Vector3.X);
+                setSpinnerState(SpinnerIdentifier.V2, Vector3.Y);
                 v1 = Vector3.X; v2 = Vector3.Y;
             }
         }
@@ -568,18 +775,47 @@ public class Grid2DPanel extends javax.swing.JPanel {
         gm.setDeleteMode(!gm.isDeleteMode());
     }//GEN-LAST:event_jCheckBox1ActionPerformed
 
+    private void queryBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_queryBtnActionPerformed
+        discoverExisting();
+    }//GEN-LAST:event_queryBtnActionPerformed
+
+    private void normalizeV1BtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_normalizeV1BtnActionPerformed
+        setSpinnerState(
+                SpinnerIdentifier.V1,
+                v1.normalize()
+        );
+    }//GEN-LAST:event_normalizeV1BtnActionPerformed
+
+    private void normalizeV2BtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_normalizeV2BtnActionPerformed
+        setSpinnerState(
+                SpinnerIdentifier.V2,
+                v2.normalize()
+        );
+    }//GEN-LAST:event_normalizeV2BtnActionPerformed
+
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JPanel btnPanel;
+    private javax.swing.JSlider directionAxesLengthSlider;
     private javax.swing.JPanel drawPanel;
     private javax.swing.JCheckBox jCheckBox1;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
     private javax.swing.JSeparator jSeparator1;
+    private javax.swing.JLabel logLbl;
+    private javax.swing.JButton normalizeV1Btn;
+    private javax.swing.JButton normalizeV2Btn;
+    private javax.swing.JSpinner originXSpinner;
+    private javax.swing.JSpinner originYSpinner;
+    private javax.swing.JSpinner originZSpinner;
     private javax.swing.JComboBox<String> presetComboBox;
-    private javax.swing.JLabel propertiesLbl;
+    private javax.swing.JButton queryBtn;
     private javax.swing.JButton renderBtn;
     private javax.swing.JButton setOriginBtn;
+    private javax.swing.JSpinner v1XSpinner;
+    private javax.swing.JSpinner v1YSpinner;
+    private javax.swing.JSpinner v1ZSpinner;
+    private javax.swing.JSpinner v2XSpinner;
+    private javax.swing.JSpinner v2YSpinner;
+    private javax.swing.JSpinner v2ZSpinner;
     private javax.swing.JButton xCompBtn;
     private javax.swing.JButton yCompBtn;
     // End of variables declaration//GEN-END:variables