+++ src/main/java/com/j3d/utility/Colors.java
@@ -1,0 +1,9 @@
+package com.j3d.utility;
+
+import java.awt.*;
+
+public abstract class Colors {
+    public static Color alphaChannel(Color original, int alpha) {
+        return new Color(original.getRed(), original.getGreen(), original.getBlue(), alpha);
+    }
+}